Models/EmbeddedModels/embedded_model_huggingface.py
from langchain_huggingface import HuggingFaceEmbeddings
from dotenv import load_dotenv

load_dotenv()

embedding = HuggingFaceEmbeddings(
    model_name = "sentence-transformers/all-MiniLM-L6-v2"
)

text = "what's your name"

vector = embedding.embed_query(text)

print(vector)