requirements.txt
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai


# Hugging Face
langchain-huggingface
transformers
huggingface-hub
sentence-transformers

python-dotenv
numpy 
scikit-learn